requirements.txt
nmap-python==1.5.4
python-nmap==0.7.1
scapy==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
jinja2==3.1.2
pyyaml==6.0.1
colorama==0.4.6
rich==13.7.0
click==8.1.7
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.4
ipaddress==1.0.23
netifaces==0.11.0
dnspython==2.4.2
cryptography==41.0.8
paramiko==3.4.0
shodan==1.30.0
validators==0.22.0
